Add support for spices
Spices are just feature flags with a fancier name. They are initialized
by the Bowl (core) and collected from the top-level command by the
CommandTreeBuilder ingredient.

The bowl gains the has_spice() method to let ingredients decide
if something should be enabled or not.

Spices are designed as a way to support feature gating for things that
might not be suitable to be on by default. Application developers can
pick spices to select the behavior they desire.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>