ingredients: Don't use print(..., flush=True)
This patch makes the code portable to Python 3.2 where flush is not yet
available as an argument to print(). There are also some unit tests to
ensure it really works as expected.

In addition, I found that calling aprint(..., file=None) (with
explicitly passed None) will misbehave. This is now fixed as well.

And another amend, I had to add funcsigs to tests_require as it is
needed by mock and setuptools doesn't pull it in correctly otherwise.

Fixes: https://github.com/zyga/guacamole/issues/9
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>