Add terminal awareness ingredient.
This rather huge patch adds the terminal awareness ingredient. This
ingredient implements terminal emulator fingerprinting coupled with a
look-aside table of features and their implementation status for each of
the multiple supported terminal emulator programs.

Extensive care was taken to support virtually everything that is
actually used out there in the field.

The list of detected, mainstream terminals includes:
 - cmd.exe on Windows
 - Apple's Terminal.app on OS X
 - Gnome Terminal
 - KDE Konsole
 - X11 xterm
 - Linux console (think vt1-6)

The list of niche terminals also supported includes:
 - iTerm on OS X
 - aterm
 - eterm
 - pterm
 - rxvt
 - terminator
 - terminology

In addition the following terminal multiplexers are supported:
 - screen
 - tmux

Lastly all remote terminals connected over SSH are identified as such
but the true nature of the terminal is not known.

This patch also contains an interesting example, examples/terminal.py.
Running this program with the "check" subcommand will create a visual
"test case" that shows all the information that guacamole managed to
find out along with a breakdown for all of the features and their
status. This can be used to validate if something works as expected.
Unfortunately it is a out-of-band knowledge so it will have to be
updated over time.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>